requirements.txt
langchain
openai
ollama
chromadb
faiss-cpu
streamlit
python-dotenv
slack-sdk
pyyaml
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-readers-file
PyMuPDF
argcomplete==3.6.2
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
slack_sdk==3.35.0
Werkzeug==3.1.3
pyigloo @ git+https://github.com/xkahn/pyigloo.git
langchain_huggingface
langchain_postgres
langchain_community
